Add reviewer changes
